[TASK] Streamline PDF page parsing in ext:indexed_search
This change streamlines the internal API at
FileContentParser::readFileContent when parsing PDF files.
Claims that it would be a "security vulnerability" are not
correct, since the internal processing relies on sanitized
data from FileContentParser::fileContentParts.

Resolves: #108644
Releases: main, 14.3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/95603
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>

+++ typo3/sysext/indexed_search/Classes/FileContentParser.php
@@ -446,13 +446,15 @@ class FileContentParser
                     $pdfInfo = $this->splitPdfInfo($res);
                     unset($res);
                     if ((int)($pdfInfo['pages'] ?? 0)) {
-                        [$low, $high] = explode('-', $cPKey);
-                        // Get pdf content:
-                        $tempFileName = GeneralUtility::tempnam('Typo3_indexer');
+                        [$low, $high] = array_map(intval(...), explode('-', $cPKey, 2) + [0, 0]);
                         // Create temporary name
-                        @unlink($tempFileName);
+                        $tempFileName = GeneralUtility::tempnam('Typo3_indexer');
                         // Delete if exists, just to be safe.
-                        $cmd = $this->app['pdftotext'] . ' -f ' . $low . ' -l ' . $high . ' -enc UTF-8 -q ' . escapeshellarg($absFile) . ' ' . $tempFileName;
+                        if (@is_file($tempFileName)) {
+                            unlink($tempFileName);
+                        }
+                        // Get pdf content:
+                        $cmd = $this->app['pdftotext'] . ' -f ' . escapeshellarg((string)$low) . ' -l ' . escapeshellarg((string)$high) . ' -enc UTF-8 -q ' . escapeshellarg($absFile) . ' ' . escapeshellarg($tempFileName);
                         CommandUtility::exec($cmd);
                         if (@is_file($tempFileName)) {
                             $content = (string)file_get_contents($tempFileName);
